♻️ Refactor: N+1문제방지,타입변환

+++ src/main/java/com/codeit/duckhu/domain/user/dto/PowerUserStatsDto.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 public record PowerUserStatsDto(
         UUID userId,
         Double reviewScoreSum,
-        Long likedCount,
-        Long commentCount
+        Integer likedCount,
+        Integer commentCount
 ) {
 }
